requirements.txt
appium-python-client==5.2.6
pytest
pytest-html
pytest-json-report
# scikit-learn
# pandas
# matplotlib
# jupyter
# shap